Split main screen into smaller modules
